# Optimized computations in commutative DGA